+++ tests/other/ResultsetTest.php
@@ -20,6 +20,16 @@
             }
         }
 
+        function testToSQL() {
+            $resultset = new Resultset([['foo' => 'bar', 'bar' => 'foo'],
+                                        ['foo' => 3, 'bar' => '']]);
+
+            $simpleExpect = 'INSERT INTO `phu`(`foo`,`bar`)
+                                 VALUES(\'bar\',\'foo\'),(3,NULL)';
+
+            $this->assertEquals($simpleExpect, $resultset->toInsertStatement('phu', 'INSERT', false));
+        }
+
         function testConstruct() {
             $this->reloadResultset();
             $this->assertEquals($this->dataArray, $this->set->toArray());
@@ -77,6 +87,22 @@
                                  ['foo' => 9, 'bar' => -9]], $this->set->toArray());
         }
 
+        function testMultiplyValue() {
+            $this->reloadResultset();
+            $this->set->multiplyValue(['bar' => 10], self::$spec);
+
+            $this->assertEquals([['foo' => 0, 'bar' => 0],
+                                 ['foo' => 1, 'bar' => -1],
+                                 ['foo' => 2, 'bar' => -2],
+                                 ['foo' => 3, 'bar' => -3],
+                                 ['foo' => 4, 'bar' => -4],
+                                 ['foo' => 5, 'bar' => -5],
+                                 ['foo' => 6, 'bar' => -60],
+                                 ['foo' => 7, 'bar' => -70],
+                                 ['foo' => 8, 'bar' => -80],
+                                 ['foo' => 9, 'bar' => -9]], $this->set->toArray());
+        }
+
         function testDeleteWhere() {
             $this->reloadResultset();
             $this->set->deleteWhere(self::$spec);
@@ -87,6 +113,17 @@
                          ['foo' => 4, 'bar' => -4],
                          ['foo' => 5, 'bar' => -5],
                          ['foo' => 9, 'bar' => -9]];
+            $actual   = array_values($this->set->toArray());
+
+            $this->assertEquals($expected, $actual);
+        }
+
+        function testKeepWhere() {
+            $this->reloadResultset();
+            $this->set->keepWhere(self::$spec);
+            $expected = [['foo' => 6, 'bar' => -6],
+                         ['foo' => 7, 'bar' => -7],
+                         ['foo' => 8, 'bar' => -8]];
             $actual   = array_values($this->set->toArray());
 
             $this->assertEquals($expected, $actual);
@@ -129,6 +166,22 @@
                                  ['foo' => 6, 'bar' => 'bb-6'],
                                  ['foo' => 7, 'bar' => 'bb-7'],
                                  ['foo' => 8, 'bar' => 'bb-8'],
+                                 ['foo' => 9, 'bar' => -9]], $this->set->toArray());
+        }
+
+        function testSetValue() {
+            $this->reloadResultset();
+            $this->set->setValue(['bar' => 'bb'], self::$spec);
+
+            $this->assertEquals([['foo' => 0, 'bar' => 0],
+                                 ['foo' => 1, 'bar' => -1],
+                                 ['foo' => 2, 'bar' => -2],
+                                 ['foo' => 3, 'bar' => -3],
+                                 ['foo' => 4, 'bar' => -4],
+                                 ['foo' => 5, 'bar' => -5],
+                                 ['foo' => 6, 'bar' => 'bb'],
+                                 ['foo' => 7, 'bar' => 'bb'],
+                                 ['foo' => 8, 'bar' => 'bb'],
                                  ['foo' => 9, 'bar' => -9]], $this->set->toArray());
         }
     }